First level and general features
Upload of the first level of the game where player could move, jump and open the door following to the second level. Player has a health bar as well that could not be however modified on this level. Pause menu is also provided though the "Restart" button is currently under development.