Major update of different features
The structure of all the levels is now costructed. Moreover, enemies appearance was updated. Furthermore, the fourth level is completely finished at the moment. In addition, player could die from this moment. Finally, the boss was added to the last location. However, its behavior has not been completely implemented yet.